+++ context/app/static/js/components/Detail/entityHeader/EntityHeaderContent/EntityHeaderContent.jsx
@@ -4,15 +4,8 @@ import { useTransition, animated } from 'react-spring';
 
 import VizualizationThemeSwitch from 'js/components/Detail/visualization/VisualizationThemeSwitch';
 import VisualizationShareButton from 'js/components/Detail/visualization/VisualizationShareButton';
-import VisualizationFullScreenButton from 'js/components/Detail/visualization/VisualizationFullScreenButton';
-import {
-  StyledDatasetIcon,
-  StyledSampleIcon,
-  StyledDonorIcon,
-  FlexContainer,
-  FullscreenToggleButton,
-  RightDiv,
-} from './style';
+import VisualizationCollapseButton from 'js/components/Detail/visualization/VisualizationCollapseButton';
+import { StyledDatasetIcon, StyledSampleIcon, StyledDonorIcon, FlexContainer, RightDiv } from './style';
 import EntityHeaderItem from '../EntityHeaderItem';
 
 const iconMap = {
@@ -42,8 +35,8 @@ function EntityHeaderContent({ display_doi, entity_type, data, shouldDisplayHead
           {vizIsFullscreen && (
             <RightDiv>
               <VisualizationShareButton />
-              <VizualizationThemeSwitch toggleButtonComponent={FullscreenToggleButton} />
-              <VisualizationFullScreenButton />
+              <VizualizationThemeSwitch />
+              <VisualizationCollapseButton />
             </RightDiv>
           )}
         </AnimatedFlexContainer>