+++ context/app/static/js/components/Header/Header/Header.jsx
@@ -2,17 +2,22 @@ import React, { useRef } from 'react';
 
 import EntityHeader from 'js/components/Detail/entityHeader/EntityHeader';
 import useVisualizationStore from 'js/stores/useVisualizationStore';
+import useEntityStore from 'js/stores/useEntityStore';
 import HeaderAppBar from '../HeaderAppBar';
 import HeaderContent from '../HeaderContent';
 
+const entityStoreSelector = (state) => state.summaryComponentObserver;
 const visualizationSelector = (state) => ({
   vizIsFullscreen: state.vizIsFullscreen,
 });
 
 function Header() {
   const anchorRef = useRef(null);
+  const { summaryInView, summaryEntry } = useEntityStore(entityStoreSelector);
   const displayEntityHeader =
-    window.location.pathname.startsWith('/browse') && !window.location.pathname.startsWith('/browse/collection');
+    summaryEntry &&
+    window.location.pathname.startsWith('/browse') &&
+    !window.location.pathname.startsWith('/browse/collection');
 
   const { vizIsFullscreen } = useVisualizationStore(visualizationSelector);
 
@@ -20,7 +25,7 @@ function Header() {
     <>
       <HeaderAppBar
         anchorRef={anchorRef}
-        elevation={displayEntityHeader ? 0 : 4}
+        elevation={!summaryInView || vizIsFullscreen ? 0 : 4}
         shouldConstrainWidth={!vizIsFullscreen}
       >
         <HeaderContent anchorRef={anchorRef} />