+++ context/app/static/js/components/Search/SearchBarLayout/style.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const Flex = styled.div`
+  display: flex;
+  width: 100%;
+`;
+
+const StyledDiv = styled.div`
+  display: flex;
+  margin-left: auto;
+  align-items: center;
+`;
+
+export { Flex, StyledDiv };