+++ context/app/static/js/components/Search/SearchWrapper.jsx
@@ -1,24 +1,15 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 
-import {
-  SearchkitManager,
-  SearchkitProvider,
-  SearchBox,
-  LayoutResults,
-  ActionBar,
-  ActionBarRow,
-  NoHits,
-  HitsStats,
-  LayoutBody,
-  Pagination,
-} from 'searchkit'; // eslint-disable-line import/no-duplicates
+import { SearchkitManager, SearchkitProvider, LayoutResults, NoHits, LayoutBody } from 'searchkit'; // eslint-disable-line import/no-duplicates
 
 import Accordions from './Accordions';
-import ResultsTable from './ResultsTable';
+import PaginationWrapper from './PaginationWrapper';
+import SearchBarLayout from './SearchBarLayout';
 import { resultFieldsToSortOptions } from './utils';
 import { StyledSideBar } from './style';
 import './Search.scss';
+import { NoResults, SearchError } from './noHitsComponents';
 
 function SearchWrapper(props) {
   const {
@@ -31,45 +22,45 @@ function SearchWrapper(props) {
     httpHeaders,
     searchUrlPath,
     queryFields,
+    type,
     isLoggedIn,
+    resultsComponent: ResultsComponent,
   } = props;
-  const sortOptions = resultFieldsToSortOptions(resultFields);
-  const resultFieldIds = resultFields.map((field) => field.id).concat(idField);
+  const [searchView, setSearchView] = useState('table');
+
+  const sortOptions = resultFieldsToSortOptions(resultFields.table);
+  const resultFieldIds = [...resultFields.table, ...resultFields.tile].map((field) => field.id).concat(idField);
   const searchkit = new SearchkitManager(apiUrl, { httpHeaders, searchUrlPath });
 
   return (
     <SearchkitProvider searchkit={searchkit}>
-      <LayoutBody>
-        <StyledSideBar>
-          <SearchBox autofocus queryFields={queryFields} />
-          <Accordions filters={filters} />
-        </StyledSideBar>
-        <LayoutResults>
-          <ActionBar>
-            <ActionBarRow>
-              <HitsStats
-                translations={{
-                  'hitstats.results_found': '{hitCount} results found',
-                }}
-              />
-            </ActionBarRow>
-          </ActionBar>
-          <ResultsTable
-            sortOptions={sortOptions}
-            hitsPerPage={hitsPerPage}
-            resultFields={resultFields}
-            detailsUrlPrefix={detailsUrlPrefix}
-            idField={idField}
-            resultFieldIds={resultFieldIds}
-          />
-          <NoHits
-            translations={{
-              'NoHits.NoResultsFound': `No results found. ${isLoggedIn ? '' : 'Login to view more results.'}`,
-            }}
-          />
-          <Pagination showNumbers />
-        </LayoutResults>
-      </LayoutBody>
+      <>
+        <SearchBarLayout
+          queryFields={queryFields}
+          searchView={searchView}
+          setSearchView={setSearchView}
+          sortOptions={sortOptions}
+        />
+        <LayoutBody>
+          <StyledSideBar>
+            <Accordions filters={filters} />
+          </StyledSideBar>
+          <LayoutResults>
+            <ResultsComponent
+              sortOptions={sortOptions}
+              hitsPerPage={hitsPerPage}
+              resultFields={resultFields[searchView]}
+              detailsUrlPrefix={detailsUrlPrefix}
+              idField={idField}
+              resultFieldIds={resultFieldIds}
+              searchView={searchView}
+              type={type}
+            />
+            <NoHits component={<NoResults isLoggedIn={isLoggedIn} />} errorComponent={SearchError} />
+            <PaginationWrapper />
+          </LayoutResults>
+        </LayoutBody>
+      </>
     </SearchkitProvider>
   );
 }
@@ -79,25 +70,37 @@ SearchWrapper.propTypes = {
   filters: PropTypes.objectOf(PropTypes.array).isRequired,
   detailsUrlPrefix: PropTypes.string.isRequired,
   idField: PropTypes.string.isRequired,
-  resultFields: PropTypes.arrayOf(
-    PropTypes.exact({
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
-      translations: PropTypes.objectOf(PropTypes.string),
-    }),
-  ).isRequired,
+  resultFields: PropTypes.exact({
+    table: PropTypes.arrayOf(
+      PropTypes.exact({
+        id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+        translations: PropTypes.objectOf(PropTypes.string),
+      }),
+    ),
+    tile: PropTypes.arrayOf(
+      PropTypes.exact({
+        id: PropTypes.string.isRequired,
+        name: PropTypes.string.isRequired,
+        translations: PropTypes.objectOf(PropTypes.string),
+      }),
+    ),
+  }).isRequired,
   hitsPerPage: PropTypes.number.isRequired,
   httpHeaders: PropTypes.objectOf(PropTypes.string),
 
   searchUrlPath: PropTypes.string,
   queryFields: PropTypes.arrayOf(PropTypes.string).isRequired,
+  type: PropTypes.string,
   isLoggedIn: PropTypes.bool,
+  resultsComponent: PropTypes.func.isRequired,
 };
 
 SearchWrapper.defaultProps = {
   searchUrlPath: '_search',
   httpHeaders: {},
   isLoggedIn: false,
+  type: undefined,
 };
 
 export default SearchWrapper;