+++ context/app/static/js/components/Search/TilesSortDropdown/index.js
@@ -1,0 +1,3 @@
+import TilesSortDropdown from './TilesSortDropdown';
+
+export default TilesSortDropdown;