+++ context/app/static/js/components/ServiceStatusTable/hooks.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+const gatewayUrl = 'https://gateway.api.hubmapconsortium.org/status.json';
+
+function useGatewayStatus() {
+  const [gatewayStatus, setGatewayStatus] = React.useState(undefined);
+  React.useEffect(() => {
+    async function getAndSetGatewayStatus() {
+      const response = await fetch(gatewayUrl);
+      if (!response.ok) {
+        console.error('Entity API status failed', response);
+        return;
+      }
+      setGatewayStatus(await response.json());
+    }
+    getAndSetGatewayStatus();
+  }, []);
+
+  return gatewayStatus;
+}
+
+export { useGatewayStatus, gatewayUrl };