+++ context/app/static/js/pages/search/DevSearch.jsx
@@ -1,11 +1,12 @@
 import React, { useContext } from 'react';
-import Typography from '@material-ui/core/Typography';
 import { ExistsQuery, BoolMustNot } from 'searchkit';
 
 import { getAuthHeader } from 'js/helpers/functions';
 import { AppContext } from 'js/components/Providers';
 import { field, listFilter, checkboxFilter } from 'js/components/Search/utils';
 import SearchWrapper from 'js/components/Search/SearchWrapper';
+import DevResults from 'js/components/Search/DevResults';
+import { SearchHeader } from './style';
 
 function DevSearch() {
   const { elasticsearchEndpoint, nexusToken } = useContext(AppContext);
@@ -22,12 +23,15 @@ function DevSearch() {
     // Search results field which will be appended to detailsUrlPrefix:
     idField: 'uuid',
     // Search results fields to display in table:
-    resultFields: [
-      field('entity_type', 'type'),
-      field('display_doi', 'ID'),
-      field('mapped_last_modified_timestamp', 'Last Modified'),
-      field('mapper_metadata.size', 'Doc Size'),
-    ],
+    resultFields: {
+      table: [
+        field('entity_type', 'type'),
+        field('display_doi', 'ID'),
+        field('mapped_last_modified_timestamp', 'Last Modified'),
+        field('mapper_metadata.size', 'Doc Size'),
+      ],
+      tile: [],
+    },
     // Default hitsPerPage is 10:
     hitsPerPage: 20,
     // Sidebar facet configuration:
@@ -43,6 +47,10 @@ function DevSearch() {
         listFilter('metadata.metadata.assay_category', 'assay_category'),
         listFilter('metadata.metadata.assay_type', 'assay_type'),
       ],
+      'Validation Errors': [
+        listFilter('mapper_metadata.validation_errors.absolute_path', 'Document Path'),
+        listFilter('mapper_metadata.validation_errors.absolute_schema_path', 'Schema Path'),
+      ],
       Booleans: [
         checkboxFilter('has_metadata', 'Has metadata?', ExistsQuery('metadata.metadata')),
         checkboxFilter('no_metadata', 'No metadata?', BoolMustNot(ExistsQuery('metadata.metadata'))),
@@ -50,6 +58,12 @@ function DevSearch() {
         checkboxFilter('no_files', 'No files?', BoolMustNot(ExistsQuery('files'))),
         checkboxFilter('has_files', 'Spatially Located (CCF)?', ExistsQuery('rui_location')),
         checkboxFilter('no_files', 'Not Spatially Located (CCF)?', BoolMustNot(ExistsQuery('rui_location'))),
+        checkboxFilter('has_errors', 'Validation Errors?', ExistsQuery('mapper_metadata.validation_errors')),
+        checkboxFilter(
+          'no_errors',
+          'No Validation Errors?',
+          BoolMustNot(ExistsQuery('mapper_metadata.validation_errors')),
+        ),
       ],
     },
     queryFields: ['everything'],
@@ -58,12 +72,12 @@ function DevSearch() {
 
   const allProps = { ...searchProps, apiUrl: elasticsearchEndpoint };
 
-  const wrappedSearch = <SearchWrapper {...allProps} />;
+  const wrappedSearch = <SearchWrapper {...allProps} resultsComponent={DevResults} />;
   return (
     <>
-      <Typography component="h1" variant="h2">
+      <SearchHeader component="h1" variant="h2">
         Dev Search
-      </Typography>
+      </SearchHeader>
       {wrappedSearch}
     </>
   );